+++ src/rules/global/general.ts
@@ -1,4 +1,4 @@
-import { breakableForEach } from "chootils/dist/loops";
+import { breakableForEach, forEach } from "chootils/dist/loops";
 import { MyTypes, PrendyStoreHelpers } from "../../declarations";
 import { get_globalUtils } from "../../helpers/prendyUtils/global";
 import { clearTimeoutSafe } from "../../helpers/utils";
@@ -7,7 +7,9 @@ export function get_globalGeneralRules<T_MyTypes extends MyTypes = MyTypes>(
   prendyAssets: T_MyTypes["Assets"],
   storeHelpers: T_MyTypes["StoreHelpers"]
 ) {
-  const { prendyOptions } = prendyAssets;
+  type AnyAnimationName = T_MyTypes["Main"]["AnyAnimationName"];
+
+  const { modelInfoByName, dollNames, prendyOptions, placeInfoByName } = prendyAssets;
   const { getRefs, getState, makeRules, setState, onNextTick } = storeHelpers;
   const { setGlobalState } = get_globalUtils(storeHelpers);
 
@@ -57,7 +59,6 @@ export function get_globalGeneralRules<T_MyTypes extends MyTypes = MyTypes>(
         const globalRefs = getRefs().global.main;
 
         if (globalState.isGamePaused) {
-          // setGlobalState({ timeMode: "menu", gameTimeSpeed: 0.25 });
           setGlobalState({ timeMode: "menu", gameTimeSpeed: prendyOptions.gameTimeSpeed * 0.1 });
         } else {
           setGlobalState({ timeMode: "game", gameTimeSpeed: prendyOptions.gameTimeSpeed });
@@ -69,24 +70,20 @@ export function get_globalGeneralRules<T_MyTypes extends MyTypes = MyTypes>(
     }),
     whenPauseKeyPressed: itemEffect({
       run() {
-        // console.log("pause key pressed");
-
         setState((state) => ({ global: { main: { isGamePaused: !state.global.main.isGamePaused } } }));
       },
       step: "input",
       check: { type: "keyboards", name: "main", prop: ["KeyP"], becomes: true },
     }),
-    whenElapsedGameTimeChanges: effect({
-      run(_diffInfo) {
-        const globalState = getState().global.main;
-        const elapsedGameTime = globalState.elapsedGameTime;
-        // log the value
-        // console.log("elapsedGameTime", elapsedGameTime);
-      },
-      check: { type: ["global"], name: ["main"], prop: ["elapsedGameTime"] },
-      step: "rendering",
-      atStepEnd: true,
-    }),
+    // whenElapsedGameTimeChanges: effect({
+    //   run(_diffInfo) {
+    //     const globalState = getState().global.main;
+    //     const elapsedGameTime = globalState.elapsedGameTime;
+    //   },
+    //   check: { type: ["global"], name: ["main"], prop: ["elapsedGameTime"] },
+    //   step: "rendering",
+    //   atStepEnd: true,
+    // }),
     whenASpeechBubbleShowsOrHides: effect({
       run(_diffInfo) {
         const speechBubblesState = getState().speechBubbles;
@@ -120,6 +117,58 @@ export function get_globalGeneralRules<T_MyTypes extends MyTypes = MyTypes>(
       check: { type: "speechBubbles", prop: ["isVisible"] },
       atStepEnd: true,
       step: "positionUi",
+    }),
+    whenGameTimeSpeedChanges: itemEffect({
+      run({ newValue: newGameTimeSpeed }) {
+        console.log("whenGameTimeSpeedChanges");
+
+        // loop through all dolls and set the animation speed
+        forEach(dollNames, (dollName) => {
+          const dollRefs = getRefs().dolls[dollName];
+          const { animWeights, modelName } = getState().dolls[dollName];
+          const animationNames = modelInfoByName[modelName].animationNames as AnyAnimationName[];
+
+          forEach(animationNames, (aniName) => {
+            const aniRef = dollRefs.aniGroupsRef[aniName];
+            if (!aniRef) return;
+            aniRef.speedRatio = newGameTimeSpeed;
+            // aniRef?.setWeightForAllAnimatables(animWeights[aniName]);
+          });
+        });
+
+        // set the state vid playback speed to gameTimeSpeed
+        //  get all the stateVid video refs
+        // get the names of all state vids
+
+        // get the current place name
+        const { nowPlaceName } = getState().global.main;
+
+        // loop all the camera names for the current place
+        const placeInfo = placeInfoByName[nowPlaceName];
+        const { cameraNames } = placeInfo;
+
+        const sliceVidState = getState().sliceVids[nowPlaceName];
+        const { stateVidId_playing, stateVidId_waiting } = sliceVidState;
+        // if (!stateVidId_playing) return;
+
+        const backdropVidRefs = getRefs().stateVids[stateVidId_playing];
+        const backdropWaitVidRefs = getRefs().stateVids[stateVidId_waiting];
+
+        backdropVidRefs.videoElement.playbackRate = newGameTimeSpeed;
+        backdropWaitVidRefs.videoElement.playbackRate = newGameTimeSpeed;
+
+        // loop all the camera names
+        // forEach(cameraNames, (cameraName) => {
+        //   const { stateVidName } = getState().cameras[cameraName];
+        //   const { videoRef } = getRefs().videos[stateVidName];
+        //   if (!videoRef) return;
+
+        //   videoRef.playbackRate = newGameTimeSpeed;
+        // }
+      },
+      check: { type: "global", prop: "gameTimeSpeed" },
+      atStepEnd: true,
+      step: "dollAnimation2",
     }),
   }));
 }